LayoutTests/css3/flexbox/undefined-min-width.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

Test to make sure that we do not assert on this code.

<div class="flexbox column" style="max-height: 0; overflow: hidden; line-height: 13px;" data-expected-height="0">
    <div style="min-height: 100%;" data-expected-height="0">This is a flex item.</div>
    <div style="flex: none;" data-expected-height="13">Inflexible</div>
</div>
